+++ packages/sprinkle-admin/app/assets/composables/useRoleUpdateApi.ts
@@ -3,45 +3,40 @@ import axios from 'axios'
 import { Severity } from '@userfrosting/sprinkle-core/interfaces'
 import type { ApiResponse, ApiErrorResponse } from '@userfrosting/sprinkle-core/interfaces'
 import { useAlertsStore } from '@userfrosting/sprinkle-core/stores'
-
-// TODO : Add validation - This composable is only used to associates the
-// permissions with the role. It should have a dedicated schema for this, plus
-// be merged with 'useRolePermissionsApi'
-// 'schema://requests/role/edit-field.yaml'
+import type { RolePermissionsRequest } from '../interfaces'
 
 /**
- * API used to update role.
+ * API used to update role permissions.
  *
- * This API is tied to the `RoleUpdateFieldAction` API, accessed at the
- * GET `/api/roles/r/{slug}/{field}` endpoint.
- *
- * This composable can be used to update {field} for a specific role.
+ * This API is tied to the `RolePermissionsAction` API, accessed at the
+ * PUT `/api/roles/r/{slug}/permissions` endpoint.
  */
 export function useRoleUpdateApi() {
     const apiLoading = ref<boolean>(false)
     const apiError = ref<ApiErrorResponse | null>(null)
 
-    async function submitRoleUpdate(slug: string, fieldName: string, formData: any) {
+    async function submitRolePermissions(slug: string, data: RolePermissionsRequest) {
         apiLoading.value = true
         apiError.value = null
 
         return axios
-            .put<ApiResponse>('/api/roles/r/' + slug + '/' + fieldName, formData)
+            .put<ApiResponse>('/api/roles/r/' + slug + '/permissions', data)
             .then((response) => {
                 useAlertsStore().push({
-                    ...{ style: Severity.Success },
-                    ...response.data
+                    ...response.data,
+                    style: Severity.Success
                 })
 
                 return response.data
             })
             .catch((err) => {
                 apiError.value = err.response.data
+                throw apiError.value
             })
             .finally(() => {
                 apiLoading.value = false
             })
     }
 
-    return { submitRoleUpdate, apiLoading, apiError }
+    return { submitRolePermissions, apiLoading, apiError }
 }